Scaled window x2. How can we resize this easier? Shall we just keep it static?
