+++ grid.hpp
@@ -7,31 +7,12 @@
 #include <vector>
 #include <SFML/Graphics.hpp>
 #include <memory>
+#include "tile.hpp"
 
-struct TileCoord
-{
-    int x;
-    int y;
-};
-
-class Tile : public std::enable_shared_from_this<Tile> {
-public:
-    Tile(int x, int y);
-    void setPosition(int x, int y);
-    TileCoord getPosition();
-    sf::RectangleShape getTile();
-    void setActivated();
-    bool isActivated();
-    std::shared_ptr<Tile> getActivated();
-private:
-    sf::RectangleShape _square;
-    TileCoord _position;
-    static Tile* _activated;
-};
 
 class Grid {
 public:
-    Grid(int xSize, int ySize, int tileSize);
+    Grid(const sf::Window & app);
     std::vector<std::vector<std::shared_ptr<Tile>>> getTiles();
     std::shared_ptr<Tile> getTile(int x, int y);
 private:
